Extract all "Refs"
